remove collect_args
