+++ wandb/sdk/launch/utils.py
@@ -115,29 +115,6 @@ def _is_valid_branch_name(work_dir, version):
     return False
 
 
-# TODO: Fix this dumb heuristic
-def _collect_args(args):
-    dict_args = {}
-    i = 0
-    while i < len(args):
-        arg = args[i]
-        if "=" in arg:
-            name, vals = arg.split("=")
-            dict_args[name.lstrip("-")] = vals
-            i += 1
-        elif (
-            arg.startswith("-")
-            and i < len(args) - 1
-            and not args[i + 1].startswith("-")
-        ):
-            dict_args[arg.lstrip("-")] = args[i + 1]
-            i += 2
-        else:
-            dict_args[arg.lstrip("-")] = None
-            i += 1
-    return dict_args
-
-
 def fetch_and_validate_project(
     uri,
     target_entity,
